chore: move expect library glue into expect.ts
